+++ docs/changelog.md
@@ -23,6 +23,8 @@ Added:
   (!363).
 - [CalibrationData.from_condition][extra.calibration.CalibrationData.from_condition]
   supports `DataCollection` objects to reference a point in time
+- [CalibrationData.from_correction][extra.calibration.CalibrationData.from_correction]
+  method to find the constants used to correct a particular run.
 - Pre-built packages will be available on PyPI for Python 3.10 - 3.13 from the
   next release, rather than only Python 3.10 (!377).
 - [Scan.group_data()][extra.components.Scan.group_data] method to make an xarray